requirements.txt
boto3==1.20.15
botocore==1.23.15
certifi==2021.10.8
charset-normalizer==2.0.8
cycler==0.11.0
fonttools==4.28.3
idna==3.3
inflection==0.5.1
jmespath==0.10.0
jq==1.2.1
kiwisolver==1.3.2
matplotlib==3.5.0
more-itertools==8.12.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
Quandl==3.7.0
requests==2.26.0
s3transfer==0.5.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
urllib3==1.26.7
